refactor: move atomicRearrange into Zcf
This is the next step of #6577 (atomic reallocations for
contracts). It adds native support in ZCF for atomicReallocate, which
makes it possible to update contracts from the helper to use the
ZCF-native approach, which is a prerequisite for removing the old
hazardous staging-based approach.

This PR only affects ZCF, though it updates types referenced from
inter-protocol contracts. This is required in order to keep the
mono-repo consistent, and has no run-time impact on the contracts,
even if changes are pushed to some live contracts even before the ZOE
change takes effect.

The sample contracts in Zoe have also been updated to use the new
API. These do not affect the interprotocol contracts. The only zoe
contract referenced by inter-protocol contracts is
scaledPriceAuthority, which is not changed here.

The new code was tested with versions of the inter-protocol contracts
updated to use the ZCF-native API. Those contracts were reverted to
their old form before merging this PR since the contracts should be
updated separately from Zoe. None of the contracts should be updated
to use the new reallocation API until this PR is running on
chain. They can be individually updated once this code is in place.