fix: refactor vatstore usage to fix test failures
Fixes #8012

There were a few direct accesses to the vatstore inside liveslots, which meant
that things that the fakestore code, which virtualizes vatstore access outside
of liveslots, missed them.  This broke some tests that were under development.
This commit relocates those accesses into the virtual reference manager, so that
they are also handled correctly if the vatstore is virtualized.