chore: patch @lerna/conventional-commits to avoid version bump for prereleases
We used to patch this library to reduce the `releaseType` level for
our packages that had not yet reached 1.0 . We then switched to an
upstream version which performed this reduction itself, except not
when doing a "prerelease". Now that we need to do a prerelease, we'd
like this reduction behavior in both arms of the conditional, so this
commit re-introduces our original patch.

closes #8242